refactor(client): move worker rpc + parse result types to app/types/
Per CLAUDE.md Rule #1, every type definition must live in app/types/ — never
inline in composables, workers, library files, or components. Previous commit
042156a co-located WorkerSuccess / WorkerFailure / WorkerResponse inside
app/lib/run-worker.ts and ParseGeoJSONResult / ParseCSVResult /
ParseShapefileResult inside each app/workers/parse-*.ts file, which violates
the rule.

Moves:
  - WorkerSuccess, WorkerFailure, WorkerResponse  -> app/types/worker.ts (new)
  - ParseGeoJSONResult, ParseCSVResult, ParseShapefileResult -> appended to
    app/types/file-upload.ts (alongside the related ParsedFile,
    SupportedFormat, GeometryType)

Re-exported all six new types from app/types/index.ts barrel. Updated
run-worker.ts, all three parse-*.ts workers, and file-parsers.ts to import
the types instead of declaring them inline.

Verified: pnpm run build (client) -> exit 0, pnpm run lint -> 0 warnings,
0 errors.
